ReportGenerator/requirements.txt
asgiref==3.8.1
chardet==5.2.0
contourpy==1.2.1
cycler==0.12.1
Django==4.2.14
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
fonttools==4.53.1
fpdf==1.7.2
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.2
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
